DS-1064: Move S3 related AWS helper classes into s3 package
