+++ xyz-hub-service/src/main/java/com/here/xyz/httpconnector/config/AwsS3Client.java
@@ -19,9 +19,9 @@
 package com.here.xyz.httpconnector.config;
 
 import com.here.xyz.httpconnector.CService;
-import com.here.xyz.util.service.aws.S3ObjectSummary;
 import com.here.xyz.util.service.aws.SecretManagerCredentialsProvider;
 import com.here.xyz.util.service.aws.s3.S3ClientHelper;
+import com.here.xyz.util.service.aws.s3.S3ObjectSummary;
 import java.net.URI;
 import java.net.URL;
 import java.time.Duration;
