+++ xyz-hub-service/src/main/java/com/here/xyz/httpconnector/config/JobS3Client.java
@@ -27,8 +27,8 @@ import com.here.xyz.httpconnector.util.jobs.Import;
 import com.here.xyz.httpconnector.util.jobs.ImportObject;
 import com.here.xyz.httpconnector.util.jobs.Job;
 import com.here.xyz.httpconnector.util.jobs.validate.Validator;
-import com.here.xyz.util.service.aws.S3ObjectSummary;
 import com.here.xyz.util.service.aws.s3.S3ClientHelper;
+import com.here.xyz.util.service.aws.s3.S3ObjectSummary;
 import io.vertx.core.Future;
 import java.io.BufferedReader;
 import java.io.EOFException;
