+++ xyz-jobs/xyz-job-steps/src/main/java/com/here/xyz/jobs/steps/inputs/Input.java
@@ -30,8 +30,9 @@ import com.here.xyz.XyzSerializable;
 import com.here.xyz.jobs.steps.Config;
 import com.here.xyz.jobs.steps.payloads.StepPayload;
 import com.here.xyz.jobs.util.S3Client;
-import com.here.xyz.util.service.aws.S3Uri;
 import com.here.xyz.util.service.Core;
+import com.here.xyz.util.service.aws.s3.S3ObjectSummary;
+import com.here.xyz.util.service.aws.s3.S3Uri;
 import java.io.IOException;
 import java.util.HashSet;
 import java.util.List;
@@ -44,8 +45,6 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ForkJoinPool;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
-import com.here.xyz.util.service.aws.S3ObjectSummary;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import software.amazon.awssdk.services.s3.model.HeadObjectResponse;
